Create 20x4display.c
Position the Text

By default, text is printed to the screen at the top row, second column. To change the position, use lcdPosition(lcd, COLUMN, ROW). On a 16×2 LCD, the rows are numbered from 0 to 1, and the columns are numbered from 0 to 15.

The following code prints “Hello, world!” to the bottom row, fourth column: